# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: Cowatch API
  description: 视频流共享平台 API 规范
  version: 0.2.0

servers:
  - url: /api/v1
    description: Next.js rewrite proxy

tags:
  - name: auth
    description: 用户认证
  - name: rooms
    description: 房间管理
  - name: users
    description: 用户管理
  - name: videos
    description: 视频源管理

paths:
  # ==================== 认证相关 ====================
  /auth/register:
    post:
      summary: 用户注册
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: 用户登录
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: 用户登出
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 登出成功

  /auth/me:
    get:
      summary: 获取当前登录用户信息
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== 房间相关 ====================
  /rooms:
    get:
      summary: 获取房间列表
      tags: [rooms]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'

    post:
      summary: 创建房间
      tags: [rooms]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /rooms/{roomCode}:
    get:
      summary: 获取房间详情
      tags: [rooms]
      security:
        - bearerAuth: []
      
      parameters:
        - name: roomCode
          in: path
          required: true
          description: 8位大写房间码
          schema:
            type: string
            pattern: '^[A-Z0-9]{8}$'
            example: "ABCD1234"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: 房间不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /rooms/{roomCode}/join:
    post:
      summary: 加入房间
      tags: [rooms]
      security:
        - bearerAuth: []
      parameters:
        - name: roomCode
          in: path
          required: true
          description: 8位大写房间码
          schema:
            type: string
            pattern: '^[A-Z0-9]{8}$'
            example: "ABCD1234"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: 房间密码（如果需要）
      responses:
        '200':
          description: 加入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 房间不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== 用户相关 ====================
  /users/me/recent-rooms:
    get:
      summary: 获取当前用户最近加入的房间
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecentRoom'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== 视频相关 ====================
  /videos/parse:
    post:
      summary: 解析视频源
      tags: [videos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseVideoRequest'
      responses:
        '200':
          description: 解析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoSource'
        '400':
          description: 无法解析视频源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证令牌

  schemas:
    # ==================== 认证相关 ====================
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 20
          example: "张三"
        password:
          type: string
          minLength: 6
          maxLength: 100
          example: "password123"
      required:
        - username
        - password

    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: "张三"
        password:
          type: string
          example: "password123"
      required:
        - username
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT 认证令牌
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      required:
        - user
        - token

    # ==================== 房间相关 ====================
    Room:
      type: object
      properties:
        id:
          type: string
          description: 房间内部 ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        code:
          type: string
          description: 8位大写房间码，用于加入房间
          pattern: '^[A-Z0-9]{8}$'
          example: "ABCD1234"
        name:
          type: string
          example: "我的观影房间"
        ownerId:
          type: string
          example: "user-456"
        ownerName:
          type: string
          example: "张三"
        isActive:
          type: boolean
          example: true
        hasPassword:
          type: boolean
          description: 房间是否需要密码
          example: false
        userCount:
          type: integer
          example: 5
        maxUsers:
          type: integer
          example: 20
        createdAt:
          type: string
          format: date-time
        currentVideo:
          $ref: '#/components/schemas/VideoSource'
        currentUserRole:
          type: string
          enum: [host, member, guest]
          description: 当前用户在此房间的角色
          example: "host"
        currentUserHasControl:
          type: boolean
          description: 当前用户是否有播放控制权限
          example: true
      required:
        - id
        - code
        - name
        - ownerId
        - isActive

    CreateRoomRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "周末观影房"
        password:
          type: string
          description: 可选的房间密码
        maxUsers:
          type: integer
          minimum: 2
          maximum: 50
          default: 20
      required:
        - name

    RecentRoom:
      type: object
      properties:
        room:
          $ref: '#/components/schemas/Room'
        lastVisited:
          type: string
          format: date-time
          example: "2024-01-19T10:30:00Z"
        lastWatchedVideoTitle:
          type: string
          example: "你的名字"
      required:
        - room
        - lastVisited

    # ==================== 用户相关 ====================
    User:
      type: object
      properties:
        id:
          type: string
          example: "user-123"
        username:
          type: string
          example: "张三"
        avatarUrl:
          type: string
          format: uri
          example: "https://api.dicebear.com/7.x/avataaars/svg?seed=user-123"
      required:
        - id
        - username

    # ==================== 视频相关 ====================
    VideoSource:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [bilibili, quark, youtube]
          example: "bilibili"
        url:
          type: string
          format: uri
          example: "https://www.bilibili.com/video/BV1xx411c7mD"
        title:
          type: string
          example: "电影标题"
        duration:
          type: integer
          description: 时长（秒）
        thumbnail:
          type: string
          format: uri
        streamUrl:
          type: string
          format: uri
          description: 实际播放的视频流地址
      required:
        - id
        - type
        - url

    ParseVideoRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: "https://www.bilibili.com/video/BV1xx411c7mD"
        type:
          type: string
          enum: [bilibili, quark, youtube]
          description: 视频源类型，如果不提供会自动识别
      required:
        - url

    VideoState:
      type: object
      description: 视频播放状态（通过 WebSocket 同步）
      properties:
        currentTime:
          type: number
          format: double
          example: 123.45
        isPlaying:
          type: boolean
        playbackRate:
          type: number
          format: double
          default: 1.0
        volume:
          type: number
          format: double
          minimum: 0
          maximum: 1

    # ==================== 错误相关 ====================
    Error:
      type: object
      properties:
        code:
          type: string
          example: "ROOM_NOT_FOUND"
        message:
          type: string
          example: "房间不存在"
      required:
        - code
        - message