# apps/api-gateway/internal/api/types.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for ParseVideoRequestType.
const (
	ParseVideoRequestTypeBilibili ParseVideoRequestType = "bilibili"
	ParseVideoRequestTypeQuark    ParseVideoRequestType = "quark"
	ParseVideoRequestTypeYoutube  ParseVideoRequestType = "youtube"
)

// Defines values for RoomCurrentUserRole.
const (
	Guest  RoomCurrentUserRole = "guest"
	Host   RoomCurrentUserRole = "host"
	Member RoomCurrentUserRole = "member"
)

// Defines values for VideoSourceType.
const (
	VideoSourceTypeBilibili VideoSourceType = "bilibili"
	VideoSourceTypeQuark    VideoSourceType = "quark"
	VideoSourceTypeYoutube  VideoSourceType = "youtube"
)

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	// Token JWT 认证令牌
	Token string `json:"token"`
	User  User   `json:"user"`
}

// CreateRoomRequest defines model for CreateRoomRequest.
type CreateRoomRequest struct {
	MaxUsers *int   `json:"maxUsers,omitempty"`
	Name     string `json:"name"`

	// Password 可选的房间密码
	Password *string `json:"password,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// ParseVideoRequest defines model for ParseVideoRequest.
type ParseVideoRequest struct {
	// Type 视频源类型，如果不提供会自动识别
	Type *ParseVideoRequestType `json:"type,omitempty"`
	Url  string                 `json:"url"`
}

// ParseVideoRequestType 视频源类型，如果不提供会自动识别
type ParseVideoRequestType string

// RecentRoom defines model for RecentRoom.
type RecentRoom struct {
	LastVisited           time.Time `json:"lastVisited"`
	LastWatchedVideoTitle *string   `json:"lastWatchedVideoTitle,omitempty"`
	Room                  Room      `json:"room"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// Room defines model for Room.
type Room struct {
	// Code 8位大写房间码，用于加入房间
	Code      string     `json:"code"`
	CreatedAt *time.Time `json:"createdAt,omitempty"`

	// CurrentUserHasControl 当前用户是否有播放控制权限
	CurrentUserHasControl *bool `json:"currentUserHasControl,omitempty"`

	// CurrentUserRole 当前用户在此房间的角色
	CurrentUserRole *RoomCurrentUserRole `json:"currentUserRole,omitempty"`
	CurrentVideo    *VideoSource         `json:"currentVideo,omitempty"`

	// HasPassword 房间是否需要密码
	HasPassword *bool `json:"hasPassword,omitempty"`

	// Id 房间内部 ID
	Id        string  `json:"id"`
	IsActive  bool    `json:"isActive"`
	MaxUsers  *int    `json:"maxUsers,omitempty"`
	Name      string  `json:"name"`
	OwnerId   string  `json:"ownerId"`
	OwnerName *string `json:"ownerName,omitempty"`
	UserCount *int    `json:"userCount,omitempty"`
}

// RoomCurrentUserRole 当前用户在此房间的角色
type RoomCurrentUserRole string

// User defines model for User.
type User struct {
	AvatarUrl *string `json:"avatarUrl,omitempty"`
	Id        string  `json:"id"`
	Username  string  `json:"username"`
}

// VideoSource defines model for VideoSource.
type VideoSource struct {
	// Duration 时长（秒）
	Duration *int   `json:"duration,omitempty"`
	Id       string `json:"id"`

	// StreamUrl 实际播放的视频流地址
	StreamUrl *string         `json:"streamUrl,omitempty"`
	Thumbnail *string         `json:"thumbnail,omitempty"`
	Title     *string         `json:"title,omitempty"`
	Type      VideoSourceType `json:"type"`
	Url       string          `json:"url"`
}

// VideoSourceType defines model for VideoSource.Type.
type VideoSourceType string

// GetRoomsParams defines parameters for GetRooms.
type GetRoomsParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// PostRoomsRoomCodeJoinJSONBody defines parameters for PostRoomsRoomCodeJoin.
type PostRoomsRoomCodeJoinJSONBody struct {
	// Password 房间密码（如果需要）
	Password *string `json:"password,omitempty"`
}

// GetUsersMeRecentRoomsParams defines parameters for GetUsersMeRecentRooms.
type GetUsersMeRecentRoomsParams struct {
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// PostAuthLoginJSONRequestBody defines body for PostAuthLogin for application/json ContentType.
type PostAuthLoginJSONRequestBody = LoginRequest

// PostAuthRegisterJSONRequestBody defines body for PostAuthRegister for application/json ContentType.
type PostAuthRegisterJSONRequestBody = RegisterRequest

// PostRoomsJSONRequestBody defines body for PostRooms for application/json ContentType.
type PostRoomsJSONRequestBody = CreateRoomRequest

// PostRoomsRoomCodeJoinJSONRequestBody defines body for PostRoomsRoomCodeJoin for application/json ContentType.
type PostRoomsRoomCodeJoinJSONRequestBody PostRoomsRoomCodeJoinJSONBody

// PostVideosParseJSONRequestBody defines body for PostVideosParse for application/json ContentType.
type PostVideosParseJSONRequestBody = ParseVideoRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// 用户登录
	// (POST /auth/login)
	PostAuthLogin(c *gin.Context)
	// 用户登出
	// (POST /auth/logout)
	PostAuthLogout(c *gin.Context)
	// 获取当前登录用户信息
	// (GET /auth/me)
	GetAuthMe(c *gin.Context)
	// 用户注册
	// (POST /auth/register)
	PostAuthRegister(c *gin.Context)
	// 获取房间列表
	// (GET /rooms)
	GetRooms(c *gin.Context, params GetRoomsParams)
	// 创建房间
	// (POST /rooms)
	PostRooms(c *gin.Context)
	// 获取房间详情
	// (GET /rooms/{roomCode})
	GetRoomsRoomCode(c *gin.Context, roomCode string)
	// 加入房间
	// (POST /rooms/{roomCode}/join)
	PostRoomsRoomCodeJoin(c *gin.Context, roomCode string)
	// 获取当前用户最近加入的房间
	// (GET /users/me/recent-rooms)
	GetUsersMeRecentRooms(c *gin.Context, params GetUsersMeRecentRoomsParams)
	// 解析视频源
	// (POST /videos/parse)
	PostVideosParse(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PostAuthLogin operation middleware
func (siw *ServerInterfaceWrapper) PostAuthLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAuthLogin(c)
}

// PostAuthLogout operation middleware
func (siw *ServerInterfaceWrapper) PostAuthLogout(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAuthLogout(c)
}

// GetAuthMe operation middleware
func (siw *ServerInterfaceWrapper) GetAuthMe(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetAuthMe(c)
}

// PostAuthRegister operation middleware
func (siw *ServerInterfaceWrapper) PostAuthRegister(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAuthRegister(c)
}

// GetRooms operation middleware
func (siw *ServerInterfaceWrapper) GetRooms(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRoomsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", c.Request.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetRooms(c, params)
}

// PostRooms operation middleware
func (siw *ServerInterfaceWrapper) PostRooms(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostRooms(c)
}

// GetRoomsRoomCode operation middleware
func (siw *ServerInterfaceWrapper) GetRoomsRoomCode(c *gin.Context) {

	var err error

	// ------------- Path parameter "roomCode" -------------
	var roomCode string

	err = runtime.BindStyledParameterWithOptions("simple", "roomCode", c.Param("roomCode"), &roomCode, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter roomCode: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetRoomsRoomCode(c, roomCode)
}

// PostRoomsRoomCodeJoin operation middleware
func (siw *ServerInterfaceWrapper) PostRoomsRoomCodeJoin(c *gin.Context) {

	var err error

	// ------------- Path parameter "roomCode" -------------
	var roomCode string

	err = runtime.BindStyledParameterWithOptions("simple", "roomCode", c.Param("roomCode"), &roomCode, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter roomCode: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostRoomsRoomCodeJoin(c, roomCode)
}

// GetUsersMeRecentRooms operation middleware
func (siw *ServerInterfaceWrapper) GetUsersMeRecentRooms(c *gin.Context) {

	var err error

	c.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUsersMeRecentRoomsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetUsersMeRecentRooms(c, params)
}

// PostVideosParse operation middleware
func (siw *ServerInterfaceWrapper) PostVideosParse(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostVideosParse(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/auth/login", wrapper.PostAuthLogin)
	router.POST(options.BaseURL+"/auth/logout", wrapper.PostAuthLogout)
	router.GET(options.BaseURL+"/auth/me", wrapper.GetAuthMe)
	router.POST(options.BaseURL+"/auth/register", wrapper.PostAuthRegister)
	router.GET(options.BaseURL+"/rooms", wrapper.GetRooms)
	router.POST(options.BaseURL+"/rooms", wrapper.PostRooms)
	router.GET(options.BaseURL+"/rooms/:roomCode", wrapper.GetRoomsRoomCode)
	router.POST(options.BaseURL+"/rooms/:roomCode/join", wrapper.PostRoomsRoomCodeJoin)
	router.GET(options.BaseURL+"/users/me/recent-rooms", wrapper.GetUsersMeRecentRooms)
	router.POST(options.BaseURL+"/videos/parse", wrapper.PostVideosParse)
}